Add a '-d' ('--dock') parameter to suppress setting override_redirect
Unfortunately this also may cause a flood of Configure events when xmobar and
the windowmanager disagree about the proper placement of the window.

For now I just removed triggering a Reposition event on Configure events. This
seems to be sufficient: we do want to request a new size when the screen
resolution (and thus width) is changed, but RRScreenChangeNotifyEvent takes
care of that.