Fixing CI error and dbus dependency issue
The Travis CI fails due to the command sh ./travis.sh src returns an
error due to the fact that travis.sh does not exist.

The fix is removing the command from travis.yml.

Versions of dbus > 0.13 are problematic when using cabal to build the project, especially when xmobar was first built without dbus, then building it again with dbus.